[ADD] adding method to lock execution if business app is running
save
