+++ Job/Job.csproj
@@ -8,10 +8,6 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <Folder Include="Controller\" />
-    </ItemGroup>
-
-    <ItemGroup>
       <Reference Include="Logger">
         <HintPath>..\Logger\bin\Debug\net8.0\Logger.dll</HintPath>
       </Reference>
